Votre message de commit ici

+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/ActifFinancierRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.ActifFinancier;
 
-public interface ActifFinancierRepository  extends JpaRepository<ActifFinancier,Integer> {
+public interface ActifFinancierRepository  extends JpaRepository<ActifFinancier, Long> {
 }
