+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/AnalyseTechniqueRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.ActifFinancier;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 
-public interface AnalyseTechniqueRepository extends JpaRepository<AnalyseTechnique,Integer> {
+public interface AnalyseTechniqueRepository extends JpaRepository<AnalyseTechnique,Long> {
 }
