+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/CertificationRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 
-public interface CertificationRepository extends JpaRepository<AnalyseTechnique,Integer> {
+public interface CertificationRepository extends JpaRepository<AnalyseTechnique,Long> {
 }
