+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/DonneesHistoriquesRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.DonneesHistoriques;
 
-public interface DonneesHistoriquesRepository extends JpaRepository<DonneesHistoriques,Integer> {
+public interface DonneesHistoriquesRepository extends JpaRepository<DonneesHistoriques,Long> {
 }
