+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/GraphiquePerformanceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.GraphiquePerformance;
 
-public interface GraphiquePerformanceRepository  extends JpaRepository<GraphiquePerformance,Integer> {
+public interface GraphiquePerformanceRepository  extends JpaRepository<GraphiquePerformance,Long> {
 }
