+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/ModeleActuarielRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.ModeleActuariel;
 
-public interface ModeleActuarielRepository extends JpaRepository<ModeleActuariel,Integer> {
+public interface ModeleActuarielRepository extends JpaRepository<ModeleActuariel,Long> {
 }
