+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/OrdreRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.Ordre;
 
-public interface OrdreRepository extends JpaRepository<Ordre,Integer> {
+public interface OrdreRepository extends JpaRepository<Ordre,Long> {
 }
