+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/ParticipationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.Participation;
 
-public interface ParticipationRepository extends JpaRepository<Participation,Integer> {
+public interface ParticipationRepository extends JpaRepository<Participation,Long> {
 }
