+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/PortefeuilleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.Portefeuille;
 
-public interface PortefeuilleRepository  extends JpaRepository<Portefeuille,Integer> {
+public interface PortefeuilleRepository  extends JpaRepository<Portefeuille,Long> {
 }
