+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/ProfilRepository.java
@@ -6,5 +6,5 @@ import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.Portefeuille;
 import tn.esprit.projetsalledemarche.Entity.Profil;
 
-public interface ProfilRepository extends JpaRepository<Profil,Integer> {
+public interface ProfilRepository extends JpaRepository<Profil,Long> {
 }
