+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/ProgressionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.Progression;
 
-public interface ProgressionRepository extends JpaRepository<Progression,Integer> {
+public interface ProgressionRepository extends JpaRepository<Progression,Long> {
 }
