+++ ProjetSalleDeMarche-main/src/main/java/tn/esprit/projetsalledemarche/Repository/SessionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import tn.esprit.projetsalledemarche.Entity.AnalyseTechnique;
 import tn.esprit.projetsalledemarche.Entity.Session;
 
-public interface SessionRepository extends JpaRepository<Session,Integer> {
+public interface SessionRepository extends JpaRepository<Session,Long> {
 }
